backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
django-cors-headers==4.3.1
requests==2.31.0
dj-database-url==2.1.0
environs==9.5.0
marshmallow==3.12.2
django-phonenumber-field==7.2.0
phonenumbers==8.13.32
djangorestframework==3.14.0
django-debug-toolbar==4.2.0
geopy==2.4.1
Pillow==10.2.0
rollbar==1.0.0